+++ tests/integration/test_create_hooked_encoder.py
