+++ tests/integration/test_d_vocab.py
