+++ tests/integration/test_hooked_sae.py
