+++ tests/integration/test_hooks.py
