+++ tests/integration/test_left_padding.py
