+++ tests/integration/test_loading_from_pretrained.py
