+++ tests/integration/test_stop_at_layer.py
